feat(compilation): Should now be able to compile files.
The RUN command is still todo, but the compile command should
work just fine. I do expect to add a build directory into the
command so that we don't need to have it local. I also
would like that directory to be chosen, so that if we're
in say, Docker, we can build outside of mount, keeping the
working directory clean.