feat(run): Added a run subcommand
At some point, we need to improve the situation
a little, and I would like a dialoguer wizard
of sorts to generate the targets since I can
imagine them getting tedious to define by hand.
As well as a "newest-file-created" add mode
which lets you pick targets for the newest create
VHDL file.